Sync with master

+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Requests/Stavke/StavkaGetMultipleRequest.cs
@@ -4,8 +4,8 @@ namespace TD.Komercijalno.Contracts.Requests.Stavke
 {
     public class StavkaGetMultipleRequest
     {
-        public int[]? VrDok { get; set; }
-        public int[]? MagacinId { get; set; }
+        public long[]? VrDok { get; set; }
+        public long[]? MagacinId { get; set; }
         public DokumentGetMultipleRequest? DokumentFilter { get; set; }
     }
 }
